Create README.md
The flex sensor is used to detect the door motion. If safety is ON, the door opening triggers and wake up the ESP, which updates the status on Firebase about security breaking. The mobile application, receives the changed status and warns a group of people via app notification or message. It also updates the cell voltage on the firebase frequently (10 mins) and the 3.3V/2200mAh battery good to go for 2 weeks.